GO-4459: Move services
