+++ core/api/server/router.go
@@ -6,12 +6,12 @@ import (
 	swaggerFiles "github.com/swaggo/files"
 	ginSwagger "github.com/swaggo/gin-swagger"
 
-	"github.com/anyproto/anytype-heart/core/api/auth"
-	"github.com/anyproto/anytype-heart/core/api/export"
-	"github.com/anyproto/anytype-heart/core/api/object"
 	"github.com/anyproto/anytype-heart/core/api/pagination"
-	"github.com/anyproto/anytype-heart/core/api/search"
-	"github.com/anyproto/anytype-heart/core/api/space"
+	"github.com/anyproto/anytype-heart/core/api/services/auth"
+	"github.com/anyproto/anytype-heart/core/api/services/export"
+	"github.com/anyproto/anytype-heart/core/api/services/object"
+	"github.com/anyproto/anytype-heart/core/api/services/search"
+	"github.com/anyproto/anytype-heart/core/api/services/space"
 )
 
 // NewRouter builds and returns a *gin.Engine with all routes configured.
